+++ src/novelyst_.py
@@ -21,6 +21,7 @@ import os
 import sys
 from pathlib import Path
 from pywriter.pywriter_globals import *
+from pywriter.config.configuration import Configuration
 from novelystlib.config.nv_configuration import NvConfiguration
 from novelystlib.view_controller.novelyst_tk import NovelystTk
 
@@ -97,9 +98,6 @@ OPTIONS = dict(
     detach_prop_win=False,
     clean_up_yw=False,
 )
-APPLICATIONS = {
-    'extension_txt': r'c:\Program Files (x86)\Zim Desktop Wiki\zim.exe',
-    }
 
 
 def main():
@@ -117,15 +115,19 @@ def main():
 
     #--- Load configuration.
     iniFile = f'{configDir}/{APPNAME}.ini'
-    configuration = NvConfiguration(SETTINGS, OPTIONS)
+    configuration = Configuration(SETTINGS, OPTIONS)
     configuration.read(iniFile)
     kwargs = {}
     kwargs.update(configuration.settings)
-    kwargs.update(configuration.applications)
     kwargs.update(configuration.options)
 
     #--- Instantiate the app object.
     app = NovelystTk('novelyst @release', tempDir, **kwargs)
+
+    #--- Launchers for opening linked non-standard filetypes.
+    launcherConfig = NvConfiguration()
+    launcherConfig.read(f'{configDir}/launchers.ini')
+    app.launchers = launcherConfig.settings
 
     #--- Load a project, if specified.
     try:
@@ -148,8 +150,6 @@ def main():
             configuration.options[keyword] = app.kwargs[keyword]
         elif keyword in configuration.settings:
             configuration.settings[keyword] = app.kwargs[keyword]
-        elif keyword in configuration.applications:
-            configuration.applications[keyword] = app.kwargs[keyword]
     configuration.write(iniFile)
 
     #--- Delete the temporary files.