+++ src/novelystlib/plugin/plugin_collection.py
@@ -59,7 +59,7 @@ class PluginCollection(dict):
         except ValueError:
             # Set defaults for testing.
             self.majorVersion = 4
-            self.minorVersion = 36
+            self.minorVersion = 43
             self.patchlevel = 0
 
     def delete_file(self, moduleName):