+++ src/novelystlib/view_controller/novelyst_tk.py
@@ -176,7 +176,10 @@ class NovelystTk(MainTk):
         self.guiStyle = ttk.Style()
 
         self.plugins = PluginCollection(self)
-        # Dict-like Container for registered plugin objects.
+        # dict-like Container for registered plugin objects
+
+        self.launchers: dict[str] = {}
+        # launchers for opening linked non-standard filetypes.
 
         self.tempDir = tempDir
         self.kwargs = kwargs
@@ -408,17 +411,6 @@ class NovelystTk(MainTk):
         self.root.bind(self._KEY_SHOW_PLANNING[0], lambda event: self.tv.show_branch(self.tv.PL_ROOT))
         self.root.bind(self._KEY_SHOW_PROJECTNOTES[0], lambda event: self.tv.show_branch(self.tv.PN_ROOT))
         self.root.bind(self._KEY_SHOW_HELP[0], lambda event: webbrowser.open(self._HELP_URL))
-
-        #--- Get applications for opening linked non-standard filetypes.
-        self.applications = {}
-        for k in kwargs:
-            if k.startswith('launch'):
-                try:
-                    __, ext = k.split('_')
-                except:
-                    pass
-                else:
-                    self.applications[f'.{ext}'] = kwargs[k]
 
     @property
     def isModified(self):