+++ docs/changelog.md
@@ -8,9 +8,10 @@
 
 See the [GitHub "Features" project](https://github.com/users/peter88213/projects/1).
 
-### v4.43.0
+### v4.43.1
 
-- A file can be linked to each character, location, and item. 
+- A file can be linked to each character, location, and item.
+- The applications for opening linked non-standard filetypes can be specified in "launchers.ini". 
 
 Based on PyWriter v12.15.3